chore: move last scale banner
