+++ src/ui/pages/app-detail-service-scale.tsx
@@ -1,22 +1,17 @@
-import { prettyDateTime } from "@app/date";
 import {
   type ServiceSizingPolicyEditProps,
-  cancelServicesOpsPoll,
   containerSizesByProfile,
   fetchApp,
   fetchService,
   fetchServiceSizingPoliciesByServiceId,
   getContainerProfileFromType,
-  getScaleTextFromOp,
   modifyServiceSizingPolicy,
-  pollServiceOperations,
   scaleService,
   selectAppById,
   selectContainerProfilesForStack,
   selectEndpointsByServiceId,
   selectEnvironmentById,
   selectManualScaleRecommendationByServiceId,
-  selectScaleDiff,
   selectServiceById,
   selectServiceSizingPolicyByServiceId,
   selectStackById,
@@ -31,9 +26,9 @@ import {
 import { appActivityUrl } from "@app/routes";
 import { DEFAULT_INSTANCE_CLASS, schema } from "@app/schema";
 import type { DeployServiceSizingPolicy, InstanceClass } from "@app/types";
-import { type SyntheticEvent, useEffect, useMemo, useState } from "react";
+import { type SyntheticEvent, useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router";
-import { usePoller, useValidator } from "../hooks";
+import { useValidator } from "../hooks";
 import {
   Banner,
   ButtonIcon,
@@ -42,6 +37,7 @@ import {
   IconChevronRight,
   IconRefresh,
   KeyValueGroup,
+  LastScaleBanner,
   ManualScaleReason,
   ServicePricingCalc,
   tokens,
@@ -841,65 +837,6 @@ const AutoscalingSection = ({
         </div>
       </form>
     </Box>
-  );
-};
-
-const LastScaleBanner = ({ serviceId }: { serviceId: string }) => {
-  const { latest, prev } = useSelector((s) =>
-    selectScaleDiff(s, { id: serviceId }),
-  );
-  const noScaleFound = latest.status === "unknown";
-  const latestComplete =
-    latest.status === "succeeded" || latest.status === "failed";
-  const action = pollServiceOperations({ id: serviceId });
-  const loader = useLoader(action);
-
-  const poller = useMemo(() => action, [serviceId]);
-  const cancel = useMemo(() => cancelServicesOpsPoll(), []);
-  usePoller({ action: poller, cancel });
-
-  const scaleTextCur = getScaleTextFromOp(latest);
-  let tail = <span className="font-bold">{scaleTextCur}</span>;
-  if (prev.id !== "") {
-    const scaleTextPrev = getScaleTextFromOp(prev);
-    if (scaleTextPrev !== "") {
-      tail = (
-        <>
-          from <span>{scaleTextPrev}</span>{" "}
-          {scaleTextCur !== "" ? (
-            <>
-              to <span className="font-bold">{scaleTextCur}</span>
-            </>
-          ) : (
-            ""
-          )}
-        </>
-      );
-    }
-  }
-
-  if (loader.isInitialLoading) {
-    return null;
-  }
-
-  if (noScaleFound) {
-    return null;
-  }
-
-  if (!latestComplete) {
-    return (
-      <Banner variant="progress">
-        <span className="font-bold">Scale in Progress</span>{" "}
-        {getScaleTextFromOp(latest)}
-      </Banner>
-    );
-  }
-
-  return (
-    <Banner variant="info">
-      <span className="font-bold">Last Scale:</span>{" "}
-      {prettyDateTime(latest.createdAt)} {tail}
-    </Banner>
   );
 };
